+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added default hotkey for inserting page breaks
 
+### Fixed
+
+- Fixed the cursor position after inserting page breaks
+
 ## 1.0.0 - 2025.10.04
 
 ### Added